Fix Run Command Args List and Captured Output
Borg removes the first value from the args list. The list being passed
in started with the command name so that was getting removed. Added
`borgapi` as the first argument for it to get removed when the command
is called now.

Used to only capture output from stdout when running a command. Some
commands output to stderr when listing json values. Annoyingly, these
commands sometime use the logger when not outputing json so the only
time we can return the values is when it is captured as a json, but
this is not always the case.