Fix arg generator not splitting lists correctly
List args were being added as the list item instead of multiple times.
This was because of a bad type check where the hints were `typing.List`
instead of `list`. Check is a bit more expensive now, so the
str / int check was moved above it and made explicit instead of the
default, if nothing matches a TypeError is raised. Only types accepted
for the options are bool, int, list, and str. If any new ones are added
this check will need to be updated (obviously).

resolves #4