Update Latest Version of Borg Required
Update to version 1.2.3 of borgbackup and fix an issue where mounting
and unmount take a couple seconds so the tests were failing. Added a
small sleep after the command to let the OS take it's time to do that.

Perhaps should look into a way to either wait for that to continue
or have a flag to wait and continue by default.